move to auto config
