+++ multitenancy-spring-boot-autoconfigure/src/main/java/org/axonframework/extensions/multitenancy/autoconfig/MultiTenantPersistentStreamMessageSource.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.axonframework.extensions.multitenancy.components.eventstore;
+package org.axonframework.extensions.multitenancy.autoconfig;
 
 
 import io.axoniq.axonserver.connector.event.PersistentStreamProperties;
@@ -22,6 +22,7 @@ import org.axonframework.config.Configuration;
 import org.axonframework.extensions.multitenancy.components.MultiTenantAwareComponent;
 import org.axonframework.extensions.multitenancy.components.TenantDescriptor;
 import org.axonframework.common.Registration;
+import org.axonframework.extensions.multitenancy.components.eventstore.MultiTenantSubscribableMessageSource;
 
 import java.util.Collections;
 import java.util.Map;
